refactor: mocks into separate file
